+++ src/Components/TriggerEditForm/Components/Form.less
@@ -1,0 +1,28 @@
+.form {
+    max-width: 800px;
+}
+
+.label {
+    flex-shrink: 0;
+    width: 150px;
+}
+
+.label-for-group {
+    align-self: flex-start;
+    padding-top: 7px;
+}
+
+.row {
+    display: flex;
+    align-items: baseline;
+    margin-bottom: 15px;
+
+    textarea {
+        vertical-align: top;
+    }
+}
+
+.control {
+    margin-left: auto;
+    width: calc(~'100% - 150px');
+}
