+++ src/Components/TriggerEditForm/TriggerEditForm.less
@@ -7,35 +7,6 @@
     top:3px
 }
 
-.form {
-    max-width: 800px;
-}
-
-.row {
-    display: flex;
-    align-items: baseline;
-    margin-bottom: 15px;
-
-    textarea {
-        vertical-align: top;
-    }
-}
-
-.label {
-    flex-shrink: 0;
-    width: 150px;
-}
-
-.control {
-    margin-left: auto;
-    width: calc(~'100% - 150px');
-}
-
-.label-for-group {
-    align-self: flex-start;
-    padding-top: 7px;
-}
-
 .target {
     top: 4px;
     position: relative;
