+++ src/Components/TriggerEditForm/TriggerEditForm.tsx
@@ -1,5 +1,6 @@
 import React, { useState, FC } from "react";
-import { ValidationWrapperV1, tooltip, ValidationInfo } from "@skbkontur/react-ui-validations";
+import { ValidationWrapperV1, tooltip } from "@skbkontur/react-ui-validations";
+import { validateRequiredString, validateTTL } from "./Validations/validations";
 import { Remarkable } from "remarkable";
 import { sanitize } from "dompurify";
 import RemoveIcon from "@skbkontur/react-icons/Remove";
@@ -9,7 +10,6 @@ import { RowStack, Fill, Fit } from "@skbkontur/react-stack-layout";
 import {
     DEFAULT_TRIGGER_TTL,
     DEFAULT_TRIGGER_TYPE,
-    LOW_TRIGGER_TTL,
     Trigger,
     TriggerSource,
     ValidateTriggerResult,
@@ -25,9 +25,10 @@ import { Status, StatusesList } from "../../Domain/Status";
 import CodeRef from "../CodeRef/CodeRef";
 import HighlightInput from "../HighlightInput/HighlightInput";
 import HelpTooltip from "../HelpTooltip/HelpTooltip";
-import EditDescriptionHelp from "./EditDescritionHelp";
-import { MetricSourceSelect } from "./MetricSourceSelect";
-import { CopyButton } from "./CopyButton";
+import EditDescriptionHelp from "./Components/EditDescritionHelp";
+import { MetricSourceSelect } from "./Components/MetricSourceSelect";
+import { CopyButton } from "./Components/CopyButton";
+import { Form, FormRow } from "./Components/Form";
 import classNames from "classnames/bind";
 
 import styles from "./TriggerEditForm.less";
@@ -36,46 +37,6 @@ const cn = classNames.bind(styles);
 
 const md = new Remarkable({ breaks: true });
 
-interface IFormProps {
-    children: React.ReactNode;
-}
-
-function Form({ children }: IFormProps): React.ReactElement {
-    return <div className={cn("form")}>{children}</div>;
-}
-
-interface IFormRowProps {
-    label?: string;
-    useTopAlignForLabel?: boolean;
-    singleLineControlGroup?: boolean;
-    style?: {
-        [key: string]: number | string;
-    };
-    children: React.ReactNode;
-}
-
-function FormRow({
-    label,
-    useTopAlignForLabel,
-    singleLineControlGroup,
-    children,
-    style,
-}: IFormRowProps): React.ReactElement {
-    return (
-        <div className={cn("row")}>
-            {label != null && (
-                <div className={cn("label", { "label-for-group": useTopAlignForLabel })}>
-                    {label}
-                </div>
-            )}
-            <div className={cn("control")}>
-                <div style={style} className={cn({ group: singleLineControlGroup })}>
-                    {children}
-                </div>
-            </div>
-        </div>
-    );
-}
 interface IProps {
     data: Partial<Trigger>;
     tags: Array<string>;
@@ -111,44 +72,6 @@ const TriggerEditForm: FC<IProps> = ({
     if (sched == null) {
         throw new Error("InvalidProgramState");
     }
-
-    const validateRequiredString = (
-        value: string,
-        message?: string
-    ): ValidationInfo | null | undefined => {
-        return value.trim().length === 0
-            ? {
-                  type: value.trim().length === 0 ? "submit" : "lostfocus",
-                  message: message || "Can't be empty",
-              }
-            : null;
-    };
-
-    const validateTTL = (value?: number): ValidationInfo | null => {
-        if (value === undefined) {
-            return {
-                type: "submit",
-                message: "Can't be empty",
-            };
-        }
-
-        if (value <= 0) {
-            return {
-                type: "lostfocus",
-                message: "Can't be zero or negative",
-            };
-        }
-
-        if (value < LOW_TRIGGER_TTL) {
-            return {
-                type: "lostfocus",
-                level: "warning",
-                message: "Low TTL can lead to false positives",
-            };
-        }
-
-        return null;
-    };
 
     const handleUpdateTarget = (targetIndex: number, value: string): void => {
         const newTargets = [...(data.targets ?? [])];
@@ -268,11 +191,10 @@ const TriggerEditForm: FC<IProps> = ({
                             {targets.length > 1 && (
                                 <Fit>
                                     <Button
+                                        icon={<RemoveIcon />}
                                         onClick={() => handleRemoveTarget(i)}
                                         data-tid="Target Remove"
-                                    >
-                                        <RemoveIcon />
-                                    </Button>
+                                    ></Button>
                                 </Fit>
                             )}
                         </RowStack>
